Add dotEnv dependency to store API keys - Organize file structure - Update tests - Create new exception for Invalid CategoryGroup handling
