+++ src/main/java/com/google/sps/enum/CategoryGroup.java
@@ -1,0 +1,7 @@
+package com.google.sps.enums;
+
+public enum CategoryGroup {
+  DINING,
+  GROCERY,
+  RECREATION;
+}
