+++ src/main/java/com/google/sps/exception/InvalidCategoryGroupException.java
@@ -1,0 +1,7 @@
+package com.google.sps.exception;
+
+public class InvalidCategoryGroupException extends Exception {
+  public InvalidCategoryGroupException(String errorMessage) {
+    super(errorMessage);
+  }
+}
