+++ src/main/java/com/google/sps/exception/InvalidDateRangeException.java
@@ -1,0 +1,7 @@
+package com.google.sps.exception;
+
+public class InvalidDateRangeException extends Exception {
+  public InvalidDateRangeException(String errorMessage) {
+    super(errorMessage);
+  }
+}
