+++ src/main/java/com/google/sps/object/Office.java
@@ -1,0 +1,29 @@
+package com.google.sps.object;
+
+public class Office {
+  String name;
+
+  // Horizontal but shown as North/South
+  double latitude;
+
+  // Verical but shown as East/West
+  double longitude;
+
+  public Office(String name, double latitude, double longitude) {
+    this.name = name;
+    this.latitude = latitude;
+    this.longitude = longitude;
+  }
+
+  public String getName() {
+    return this.name;
+  }
+
+  public double getLatitude() {
+    return this.latitude;
+  }
+
+  public double getLongitude() {
+    return this.longitude;
+  }
+}
