+++ src/main/java/com/google/sps/servlets/BusLocations.java
@@ -20,11 +20,11 @@ import com.google.appengine.api.datastore.Entity;
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
-import com.google.sps.data.CoordinateCalculator;
-import com.google.sps.data.EntityType;
-import com.google.sps.data.Office;
-import com.google.sps.data.OfficeManager;
-import com.google.sps.data.QueryHelper;
+import com.google.sps.util.CoordinateCalculator;
+import com.google.sps.enums.EntityType;
+import com.google.sps.object.Office;
+import com.google.sps.util.OfficeManager;
+import com.google.sps.util.QueryHelper;
 import java.io.IOException;
 import java.util.List;
 import java.util.stream.Collectors;
