+++ src/main/java/com/google/sps/servlets/LocationsServlet.java
@@ -23,12 +23,12 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.reflect.TypeToken;
-import com.google.sps.data.CoordinateCalculator;
-import com.google.sps.data.EntityType;
-import com.google.sps.data.InvalidDateRangeException;
-import com.google.sps.data.Office;
-import com.google.sps.data.OfficeManager;
-import com.google.sps.data.QueryHelper;
+import com.google.sps.util.CoordinateCalculator;
+import com.google.sps.enums.EntityType;
+import com.google.sps.exception.InvalidDateRangeException;
+import com.google.sps.object.Office;
+import com.google.sps.util.OfficeManager;
+import com.google.sps.util.QueryHelper;
 import java.io.IOException;
 import java.lang.reflect.Type;
 import java.time.Instant;
