--- src/main/java/com/google/sps/servlets/data/InvalidDateRangeException.java
@@ -1,7 +1,0 @@
-package com.google.sps.data;
-
-public class InvalidDateRangeException extends Exception {
-  public InvalidDateRangeException(String errorMessage) {
-    super(errorMessage);
-  }
-}
