--- src/main/java/com/google/sps/servlets/data/OfficeManager.java
@@ -1,13 +1,0 @@
-package com.google.sps.data;
-
-import java.util.HashMap;
-
-public final class OfficeManager {
-  public static final HashMap<String, Office> offices =
-      new HashMap<String, Office>() {
-        {
-          put("svl", new Office("Sunnyvale", 37.4030, -122.0326));
-          put("mtv", new Office("Mountain View", 37.4220, -122.0841));
-        }
-      };
-}
