--- src/main/java/com/google/sps/servlets/data/util/CategoryGroup.java
@@ -1,7 +1,0 @@
-package com.google.sps.data.util;
-
-public enum CategoryGroup {
-  DINING,
-  GROCERY,
-  RECREATION;
-}
